+++ src/main/java/com/doublew2w/sbs/mybatis/binding/MapperProxyFactory.java
@@ -2,7 +2,10 @@ package com.doublew2w.sbs.mybatis.binding;
 
 import com.doublew2w.sbs.mybatis.session.SqlSession;
 
+import java.lang.reflect.Method;
 import java.lang.reflect.Proxy;
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
 
 /**
  * 映射器代理工厂：负责创建映射器代理类
@@ -15,6 +18,8 @@ public class MapperProxyFactory<T> {
   /** 映射器接口类型 */
   private final Class<T> mapperInterface;
 
+  private Map<Method, MapperMethod> methodCache = new ConcurrentHashMap<>();
+
   public MapperProxyFactory(Class<T> mapperInterface) {
     this.mapperInterface = mapperInterface;
   }
@@ -26,7 +31,7 @@ public class MapperProxyFactory<T> {
    * @return 映射器代理类
    */
   public T newInstance(SqlSession sqlSession) {
-    final MapperProxy<T> mapperProxy = new MapperProxy<>(sqlSession, mapperInterface);
+    final MapperProxy<T> mapperProxy = new MapperProxy<>(sqlSession, mapperInterface, methodCache);
     return mapperInterface.cast(
         Proxy.newProxyInstance(
             mapperInterface.getClassLoader(), new Class[] {mapperInterface}, mapperProxy));