+++ src/main/java/com/doublew2w/sbs/mybatis/binding/MapperRegistry.java
@@ -1,6 +1,7 @@
 package com.doublew2w.sbs.mybatis.binding;
 
 import cn.hutool.core.lang.ClassScanner;
+import com.doublew2w.sbs.mybatis.session.Configuration;
 import com.doublew2w.sbs.mybatis.session.SqlSession;
 
 import java.util.HashMap;
@@ -15,9 +16,15 @@ import java.util.Set;
  * @project: sbs-mybatis
  */
 public class MapperRegistry {
+  /** 配置类 */
+  private final Configuration config;
 
   /** 将已添加的映射器代理加入到 HashMap */
   private final Map<Class<?>, MapperProxyFactory<?>> knownMappers = new HashMap<>();
+
+  public MapperRegistry(Configuration config) {
+    this.config = config;
+  }
 
   public <T> T getMapper(Class<T> type, SqlSession sqlSession) {
     final MapperProxyFactory<T> mapperProxyFactory = (MapperProxyFactory<T>) knownMappers.get(type);