+++ src/main/java/com/doublew2w/sbs/mybatis/builder/BaseBuilder.java
@@ -1,0 +1,22 @@
+package com.doublew2w.sbs.mybatis.builder;
+
+import com.doublew2w.sbs.mybatis.session.Configuration;
+
+/**
+ * 建造者模式：建造者基类
+ *
+ * @author: DoubleW2w
+ * @date: 2024/9/1 16:33
+ * @project: sbs-mybatis
+ */
+public abstract class BaseBuilder {
+  protected final Configuration configuration;
+
+  public BaseBuilder(Configuration configuration) {
+    this.configuration = configuration;
+  }
+
+  public Configuration getConfiguration() {
+    return configuration;
+  }
+}