+++ src/main/java/com/doublew2w/sbs/mybatis/session/defaults/DefaultSqlSession.java
@@ -2,6 +2,8 @@ package com.doublew2w.sbs.mybatis.session.defaults;
 
 import com.alibaba.fastjson2.JSON;
 import com.doublew2w.sbs.mybatis.binding.MapperRegistry;
+import com.doublew2w.sbs.mybatis.mapping.MappedStatement;
+import com.doublew2w.sbs.mybatis.session.Configuration;
 import com.doublew2w.sbs.mybatis.session.SqlSession;
 
 /**
@@ -10,11 +12,10 @@ import com.doublew2w.sbs.mybatis.session.SqlSession;
  * @project: sbs-mybatis
  */
 public class DefaultSqlSession implements SqlSession {
-  /** 映射器注册机 */
-  private MapperRegistry mapperRegistry;
+  private final Configuration configuration;
 
-  public DefaultSqlSession(MapperRegistry mapperRegistry) {
-    this.mapperRegistry = mapperRegistry;
+  public DefaultSqlSession(Configuration configuration) {
+    this.configuration = configuration;
   }
 
   @Override
@@ -24,11 +25,24 @@ public class DefaultSqlSession implements SqlSession {
 
   @Override
   public <T> T selectOne(String statement, Object parameter) {
-    return (T) ("你被代理了！" + "方法：" + statement + " 入参：" + JSON.toJSON(parameter));
+    MappedStatement mappedStatement = configuration.getMappedStatement(statement);
+    return (T)
+        ("你被代理了！"
+            + "\n方法："
+            + statement
+            + "\n入参："
+            + JSON.toJSON(parameter)
+            + "\n待执行SQL："
+            + mappedStatement.getSql());
   }
 
   @Override
   public <T> T getMapper(Class<T> type) {
-    return mapperRegistry.getMapper(type, this);
+    return configuration.getMapper(type, this);
+  }
+
+  @Override
+  public Configuration getConfiguration() {
+    return configuration;
   }
 }