+++ src/test/java/com/doublew2w/sbs/mybatis/test/dao/IUserDao.java
@@ -1,4 +1,4 @@
-package com.doublew2w.sbs.mybatis.binding.test.dao;
+package com.doublew2w.sbs.mybatis.test.dao;
 
 /**
  * @author: DoubleW2w
@@ -9,4 +9,6 @@ public interface IUserDao {
   String queryUserName(String uId);
 
   Integer queryUserAge(String uId);
+
+  String queryUserInfoById(String uId);
 }